fix!: Expose sizeInBytes as a 64-bit BigInt scalar
GraphQL fixes Int at 32 bits, so any file over ~2.1 GB failed to
serialize: the field resolved to null and the response carried a
per-field error, degrading a whole page of results to a partial one.
Every ENCODE .hic file (6-51 GB) and the larger 4DN mcools hit this, so
it is the common case for contact maps rather than an edge case.

BigInt serializes as a JSON number rather than a string so sizeInBytes
stays usable in client arithmetic with no parsing step. The usual
objection to that — values above 2^53-1 lose precision in JavaScript —
does not bind, because 2^53 bytes is roughly 9 PB.

The input filter carries the same scalar. Widening only the output
would leave exactly the files it newly exposes unfilterable by size.

The override is keyed by (model, field name) rather than by annotation
so that widening one model's int does not silently widen every other
int in the schema.

BREAKING CHANGE: sizeInBytes is typed BigInt instead of Int on both
FileMetadataType and FileMetadataInput. A query declaring a variable as
Int for that argument now fails validation and must declare BigInt, and
generated clients must be regenerated. Clients that only read
sizeInBytes out of the response are unaffected — the wire form is still
a JSON number.

+++ src/cfdb/api/gql/inputs.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 import strawberry
 
+from cfdb.api.gql.types import BigInt
+
 
 @strawberry.input
 class AnatomyInput:
@@ -243,7 +245,10 @@ class FileMetadataInput:
     project_local_id: list[str] | None = None
     persistent_id: list[str] | None = None
     creation_time: list[str] | None = None
-    size_in_bytes: list[int] | None = None
+    # ``BigInt``, not ``int``: a 32-bit ``Int`` filter cannot name the size of
+    # a file over ~2 GB, which would leave exactly the files the widened
+    # output field exposes unfilterable.
+    size_in_bytes: list[BigInt] | None = None
     sha256: list[str] | None = None
     md5: list[str] | None = None
     filename: list[str] | None = None
